Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,6 +4,6 @@
 int main() {
     int a ,b ,c;
     scanf("%i %i %i", &a, &b, &c);
-    double d = (a + b + c)/3;
-    printf("Average: %.2lf", d);
+    float d = (a + b + c)/3;
+    printf("Average: %.2f", d);
 }
